+++ packages/app-headless-cms/src/admin/components/ContentEntryForm/useContentEntryForm.ts
@@ -50,7 +50,7 @@ interface UseContentEntryForm {
     loading: boolean;
     setLoading: Dispatch<SetStateAction<boolean>>;
     onChange: FormOnSubmit;
-    onSubmit: (data: Record<string, any>) => void;
+    onSubmit: FormOnSubmit;
     invalidFields: Record<string, string>;
     renderPlugins: CmsEditorFieldRendererPlugin[];
 }
@@ -254,6 +254,13 @@ export function useContentEntryForm(params: UseContentEntryFormParams): UseConte
         [contentModel.modelId, listQueryVariables]
     );
 
+    const onChange: FormOnSubmit = (data, form) => {
+        if (!params.onChange) {
+            return;
+        }
+        return params.onChange(data, form);
+    };
+
     const onSubmit = async (data: Record<string, any>) => {
         const fieldsIds = contentModel.fields.map(item => item.fieldId);
         const formData = pick(data, [...fieldsIds]);
@@ -272,7 +279,7 @@ export function useContentEntryForm(params: UseContentEntryFormParams): UseConte
         return createContentFrom(entry.id, gqlData);
     };
 
-    const getDefaultValues = (overrides: Record<string, any> = {}): Record<string, any> => {
+    const defaultValues = useMemo((): Record<string, any> => {
         const values: Record<string, any> = {};
         /**
          * Assign the default values:
@@ -324,20 +331,17 @@ export function useContentEntryForm(params: UseContentEntryFormParams): UseConte
                     return convertDefaultValue(field, value);
                 });
         }
-        return { ...values, ...overrides };
-    };
+        return values;
+    }, [contentModel.modelId]);
+
     return {
         /**
          * If entry is not set or entry.id does not exist, it means that form is for the new entry, so fetch default values.
          */
-        data: entry && entry.id ? entry : getDefaultValues(),
+        data: entry && entry.id ? entry : defaultValues,
         loading,
         setLoading,
-        onChange:
-            params.onChange ||
-            (() => {
-                return void 0;
-            }),
+        onChange,
         onSubmit,
         invalidFields,
         renderPlugins