+++ packages/cli/commands/wcp/utils/getUser.js
@@ -1,11 +1,13 @@
-const { localStorage, log } = require("@webiny/cli/utils");
+const { log } = require("@webiny/cli/utils");
 const { request } = require("graphql-request");
+const { getWcpPat } = require("./getWcpPat");
 
 const GET_CURRENT_USER = /* GraphQL */ `
     query GetUser {
         users {
             getCurrentUser {
                 id
+                email
                 firstName
                 lastName
             }
@@ -49,7 +51,7 @@ module.exports.getUser = async () => {
         return user;
     }
 
-    const pat = localStorage().get("wcpPat");
+    const pat = getWcpPat();
     if (!pat) {
         throw new Error(
             `It seems you are not logged in. Please login using the ${log.error.hl(
@@ -59,29 +61,33 @@ module.exports.getUser = async () => {
     }
 
     try {
-        const { WCP_API_URL } = require(".");
+        const { WCP_GRAPHQL_API_URL } = require(".");
         const headers = { authorization: pat };
-        user = await request(WCP_API_URL, GET_CURRENT_USER, {}, headers).then(async response => {
-            const user = response.users.getCurrentUser;
+        user = await request(WCP_GRAPHQL_API_URL, GET_CURRENT_USER, {}, headers).then(
+            async response => {
+                const user = response.users.getCurrentUser;
 
-            const orgs = await request(WCP_API_URL, LIST_ORGS, {}, headers).then(async response => {
-                const orgs = response.orgs.listOrgs.data;
-                for (let i = 0; i < orgs.length; i++) {
-                    const org = orgs[i];
-                    org.projects = await request(
-                        WCP_API_URL,
-                        LIST_PROJECTS,
-                        { orgId: org.id },
-                        headers
-                    ).then(response => response.projects.listProjects.data);
-                }
-                return orgs;
-            });
+                const orgs = await request(WCP_GRAPHQL_API_URL, LIST_ORGS, {}, headers).then(
+                    async response => {
+                        const orgs = response.orgs.listOrgs.data;
+                        for (let i = 0; i < orgs.length; i++) {
+                            const org = orgs[i];
+                            org.projects = await request(
+                                WCP_GRAPHQL_API_URL,
+                                LIST_PROJECTS,
+                                { orgId: org.id },
+                                headers
+                            ).then(response => response.projects.listProjects.data);
+                        }
+                        return orgs;
+                    }
+                );
 
-            const projects = orgs.map(org => org.projects).flat();
+                const projects = orgs.map(org => org.projects).flat();
 
-            return { ...user, orgs, projects };
-        });
+                return { ...user, orgs, projects };
+            }
+        );
     } catch {
         throw new Error(
             `It seems the personal access token is incorrect or does not exist. Please log out and again log in using the ${log.error.hl(