+++ packages/cli/commands/wcp/utils/setWcpPat.js
@@ -1,0 +1,5 @@
+const { localStorage } = require("@webiny/cli/utils");
+
+module.exports.setWcpPat = wcpPat => {
+    localStorage().set("wcpPat", wcpPat);
+};